+++ javascripts/core/nanites/upgrades.js
@@ -76,20 +76,20 @@ function getTotalNaniteUpMult(tier) {
 }
 
 function updateUINaniteUps() {
-	document.getElementById("nanitebtn").style.display = (player.unlocked.naniteUps || player.unlocked.meltdown) ? "inline-block" : "none";
+	document.getElementById("nanites_navibtn").style.display = player.unlocked.naniteUps || player.unlocked.meltdown ? "inline-block" : "none";
 	
 	document.getElementById("nanites").innerText = notation(player.nanites.nanites);
 	
-	document.getElementById("naniteupcost0").innerText = (!player.nanites.effUpCost.eq(1)) ? (player.nanites.effUpCost + " Nanites") : "1 Nanite";
-	document.getElementById("naniteup0").className = (canBuyNaniteUp(0)) ? "naniteupbuy" : "naniteuplocked";
+	document.getElementById("nanite_upcost0").innerText = player.nanites.effUpCost.neq(1) ? (player.nanites.effUpCost + " Nanites") : "1 Nanite";
+	document.getElementById("nanite_up0").className = canBuyNaniteUp(0) ? "naniteupbuy" : "naniteuplocked";
 	
-	document.getElementById("naniteupformula12v1").style.display =  (player.nucleosynthesis <= 13) ? "inline-block" : "none";
-	document.getElementById("naniteupformula12v2").style.display =  (player.nucleosynthesis > 13) ? "inline-block" : "none";
+	document.getElementById("nanite_upformula12v1").style.display =  player.nucleosynthesis <= 13 ? "inline-block" : "none";
+	document.getElementById("nanite_upformula12v2").style.display =  player.nucleosynthesis > 13 ? "inline-block" : "none";
 
 	for (let i = 0; i < naniteUpList.length; i++) {
 		if (naniteUpList[i] != 0 && naniteUpList[i] != 31) {
-			document.getElementById("naniteupmult" + naniteUpList[i]).innerText = notation(getNaniteUpMult(naniteUpList[i])[1]);
+			document.getElementById("nanite_upmult" + naniteUpList[i]).innerText = notation(getNaniteUpMult(naniteUpList[i])[1]);
 		}
-		document.getElementById("naniteup" + naniteUpList[i]).className = (player.nanites.ups[naniteUpList[i]] == 1) ? "naniteupbought" : (canBuyNaniteUp(naniteUpList[i])) ? "naniteupbuy" : "naniteuplocked";
+		document.getElementById("nanite_up" + naniteUpList[i]).className = player.nanites.ups[naniteUpList[i]] == 1 ? "naniteupbought" : canBuyNaniteUp(naniteUpList[i]) ? "naniteupbuy" : "naniteuplocked";
 	}
 }